+++ Nemesis-Mod/src/Hardware/GPIO/DigitalPin.h
@@ -6,7 +6,7 @@
 
 class DigitalPin : public Pin {
     public:
-        DigitalPin::DigitalPin(uint32_t id);
+        DigitalPin::DigitalPin(uint8_t id);
 
         int read();
         void write(uint32_t value);
